refactor: codingzone 도메인 관련 DTO 파일 이동
