+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/response/DeleteAllInfResponseDto.java
@@ -1,4 +1,5 @@
-package com.icehufs.icebreaker.dto.response.codingzone;
+package com.icehufs.icebreaker.domain.codingzone.dto.response;
+
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,22 +8,18 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class DeleteClassResponseDto extends ResponseDto{
-        private DeleteClassResponseDto(){
+public class DeleteAllInfResponseDto extends ResponseDto{
+
+        private DeleteAllInfResponseDto(){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
 
-        public static ResponseEntity<DeleteClassResponseDto> success(){
-            DeleteClassResponseDto result = new DeleteClassResponseDto();
+        public static ResponseEntity<DeleteAllInfResponseDto> success(){
+            DeleteAllInfResponseDto result = new DeleteAllInfResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
         public static ResponseEntity<ResponseDto> notExistUser (){
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
-
-    public static ResponseEntity<ResponseDto> noExistArticle(){
-        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); }
-    
 }
