+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/response/DeleteClassResponseDto.java
@@ -1,0 +1,28 @@
+package com.icehufs.icebreaker.domain.codingzone.dto.response;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import com.icehufs.icebreaker.common.ResponseCode;
+import com.icehufs.icebreaker.common.ResponseMessage;
+import com.icehufs.icebreaker.dto.response.ResponseDto;
+
+public class DeleteClassResponseDto extends ResponseDto{
+        private DeleteClassResponseDto(){
+        super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
+    }
+
+        public static ResponseEntity<DeleteClassResponseDto> success(){
+            DeleteClassResponseDto result = new DeleteClassResponseDto();
+        return ResponseEntity.status(HttpStatus.OK).body(result);
+    }
+        public static ResponseEntity<ResponseDto> notExistUser (){
+        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
+    }
+
+    public static ResponseEntity<ResponseDto> noExistArticle(){
+        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); }
+    
+}
