+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/response/GetCodingZoneAssitantListResponseDto.java
@@ -1,11 +1,11 @@
-package com.icehufs.icebreaker.dto.response.codingzone;
+package com.icehufs.icebreaker.domain.codingzone.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
-import com.icehufs.icebreaker.dto.object.CodingZoneAssitantListItem;
+import com.icehufs.icebreaker.domain.codingzone.domain.vo.CodingZoneAssitantListItem;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 import com.icehufs.icebreaker.entity.User;
