+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/response/GetListOfGroupInfResponseDto.java
@@ -1,10 +1,10 @@
-package com.icehufs.icebreaker.dto.response.codingzone;
+package com.icehufs.icebreaker.domain.codingzone.dto.response;
 
 import com.icehufs.icebreaker.dto.response.ResponseDto;
-import com.icehufs.icebreaker.entity.GroupInfEntity;
+import com.icehufs.icebreaker.domain.codingzone.domain.entity.GroupInfEntity;
 import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
-import com.icehufs.icebreaker.dto.object.GroupInfListItem;
+import com.icehufs.icebreaker.domain.codingzone.domain.vo.GroupInfListItem;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
