+++ backend/src/main/java/com/icehufs/icebreaker/domain/codingzone/dto/response/PutAttendanceResponseDto.java
@@ -1,5 +1,4 @@
-package com.icehufs.icebreaker.dto.response.codingzone;
-
+package com.icehufs.icebreaker.domain.codingzone.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,18 +7,22 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class DeleteAllInfResponseDto extends ResponseDto{
+import lombok.Getter;
 
-        private DeleteAllInfResponseDto(){
+@Getter
+public class PutAttendanceResponseDto extends ResponseDto{
+        private PutAttendanceResponseDto() {
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
-
-        public static ResponseEntity<DeleteAllInfResponseDto> success(){
-            DeleteAllInfResponseDto result = new DeleteAllInfResponseDto();
+    
+    public static ResponseEntity<PutAttendanceResponseDto> success() {
+        PutAttendanceResponseDto result = new PutAttendanceResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
-        public static ResponseEntity<ResponseDto> notExistUser (){
+
+    public static ResponseEntity<ResponseDto> notExistUser (){
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
+    
 }
